fix(mcp): expose all 7 tools using composite provider pattern (#55)
