+++ CHANGELOG.md
@@ -10,8 +10,15 @@
   - `cupertino list-frameworks` - List available frameworks with document counts
 - All new commands support `--format text|json|markdown` output
 
+### Fixed
+- **MCP Server Tool Registration** - Fixed bug where only sample code tools were exposed (#55)
+  - Unified `CupertinoSearchToolProvider` and `SampleToolProvider` into single `CupertinoUnifiedToolProvider`
+  - All 7 MCP tools now available: `search_docs`, `list_frameworks`, `read_document`, `search_samples`, `list_samples`, `read_sample`, `read_sample_file`
+  - Updated serve command documentation to include sample code tools
+
 ### Related Issues
 - Closes #51
+- Fixes #55
 
 ---
