+++ Main.xcworkspace/contents.xcworkspacedata
@@ -5,6 +5,9 @@
       location = "group:README.md">
    </FileRef>
    <FileRef
+      location = "group:AGENTS.md">
+   </FileRef>
+   <FileRef
       location = "group:docs">
    </FileRef>
    <FileRef
