+++ Packages/Sources/SearchToolProvider/CompositeToolProvider.swift
@@ -1,0 +1,78 @@
+import Foundation
+import MCP
+import SampleIndex
+import Search
+import Shared
+
+// MARK: - Unified Cupertino Tool Provider
+
+/// Composite tool provider that delegates to both documentation and sample code providers
+public actor CompositeToolProvider: ToolProvider {
+    private let documentationTools: DocumentationToolProvider?
+    private let sampleCodeTools: SampleCodeToolProvider?
+
+    public init(searchIndex: Search.Index?, sampleDatabase: SampleIndex.Database?) {
+        self.documentationTools = searchIndex.map { DocumentationToolProvider(searchIndex: $0) }
+        self.sampleCodeTools = sampleDatabase.map { SampleCodeToolProvider(database: $0) }
+    }
+
+    // MARK: - ToolProvider
+
+    public func listTools(cursor: String?) async throws -> ListToolsResult {
+        var allTools: [Tool] = []
+
+        // Get tools from documentation provider
+        if let documentationTools {
+            let docTools = try await documentationTools.listTools(cursor: cursor)
+            allTools.append(contentsOf: docTools.tools)
+        }
+
+        // Get tools from sample code provider
+        if let sampleCodeTools {
+            let sampleTools = try await sampleCodeTools.listTools(cursor: cursor)
+            allTools.append(contentsOf: sampleTools.tools)
+        }
+
+        return ListToolsResult(tools: allTools)
+    }
+
+    public func callTool(name: String, arguments: [String: AnyCodable]?) async throws -> CallToolResult {
+        // Try documentation provider first
+        if let documentationTools {
+            let docTools = try await documentationTools.listTools(cursor: nil)
+            if docTools.tools.contains(where: { $0.name == name }) {
+                return try await documentationTools.callTool(name: name, arguments: arguments)
+            }
+        }
+
+        // Try sample code provider next
+        if let sampleCodeTools {
+            let sampleTools = try await sampleCodeTools.listTools(cursor: nil)
+            if sampleTools.tools.contains(where: { $0.name == name }) {
+                return try await sampleCodeTools.callTool(name: name, arguments: arguments)
+            }
+        }
+
+        // Tool not found in any provider
+        throw UnifiedToolError.unknownTool(name)
+    }
+}
+
+// MARK: - Tool Errors
+
+enum UnifiedToolError: Error, LocalizedError {
+    case unknownTool(String)
+    case missingArgument(String)
+    case invalidArgument(String, String)
+
+    var errorDescription: String? {
+        switch self {
+        case .unknownTool(let name):
+            return "Unknown tool: \(name)"
+        case .missingArgument(let arg):
+            return "Missing required argument: \(arg)"
+        case .invalidArgument(let arg, let reason):
+            return "Invalid argument '\(arg)': \(reason)"
+        }
+    }
+}
