--- Packages/Sources/SearchToolProvider/SearchToolProvider.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-/// Namespace for search tool providers
-public enum SearchToolProvider {
-    // Namespace root - types defined in extensions
-}
